Added compilation for windows for native codecs Modified logging to include date Resolved recorder bug

+++ codecs/nativeilbc/native/src/main/native/webrtc-ilbc/webrtc/common_audio/signal_processing/include/spl_inl.h
@@ -39,7 +39,7 @@ static __inline int16_t WebRtcSpl_SubSatW16(int16_t var1, int16_t var2) {
 }
 
 static __inline int16_t WebRtcSpl_GetSizeInBits(uint32_t n) {
-  int bits;
+  int16_t bits;
 
   if (0xFFFF0000 & n) {
     bits = 16;
